Updated and refactored movie controller, movie service, movie repository, user controller, user service, user repository, added custom exceptions

+++ CinemaApp.Domain/Services/MovieService/MovieService.cs
@@ -1,6 +1,8 @@
 ﻿using CinemaApp.DAL.Repositories.MovieRepository;
 using CinemaApp.Database.Entities.MovieModels;
+using CinemaApp.Domain.DTO;
 using CinemaApp.Domain.DTO.UserDTO;
+using CinemaApp.Domain.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,52 +21,90 @@ namespace CinemaApp.Domain.Services.MovieService
 
         public void AddMovie(Movie movie)
         {
-            //check for errors
+            //check if object fields are correct
             _movieRepository.AddMovie(movie);
+        }
+
+        public void AddSampleMovies()
+        {
+            var movies = _movieRepository.GetAllMovies().ToList();
+            if (movies.Count == 0)
+                _movieRepository.AddSampleMovies();
+            else
+                throw new ListIsNotEmptyException();
         }
 
         public void DeleteMovieById(int id)
         {
-            //check for errors
-            _movieRepository.DeleteMovieById(id);
+            var movie = _movieRepository.GetEntityById(id);
+            if (movie == null)
+                throw new ItemDoesntExistException();
+            else
+                _movieRepository.DeleteMovieById(id);
         }
 
         public IEnumerable<Movie> GetAllMovies()
         {
             var movies = _movieRepository.GetAllMovies();
-            //check for errors
+            if (movies == null)
+                throw new ListIsEmptyException();
+
             return movies;
         }
 
         public Movie GetEntityById(int id)
         {
             var movie = _movieRepository.GetEntityById(id);
-            //check for errors
+            if (movie == null)
+                throw new ItemDoesntExistException();
+
             return movie;
         }
 
         public IEnumerable<Movie> GetFivemovies()
         {
-            var movies = _movieRepository.GetFivemovies();
-            return movies;
+            var movies = _movieRepository.GetAllMovies().ToList();
+            if (movies == null)
+                throw new ListIsEmptyException();
+            if (movies.Count < 5)
+                throw new NotEnoughMoviesException();
+            return _movieRepository.GetFiveMovies();
         }
 
         public Movie GetRandomMovie()
         {
-            //check for errors
+            var movies = _movieRepository.GetAllMovies().ToList();
+            if (movies.Count == 0)
+                throw new ListIsEmptyException();
+
             return _movieRepository.GetRandomMovie();
         }
 
-        public WeeklyDiscountMovieDTO GetWeeklyDiscountMovie()
+        public DiscountDTO GetWeeklyDiscountMovie()
         {
-            //check for errors
             var weeklyDiscount = _movieRepository.GetWeeklyDiscountMovie();
-            var weeklyDiscountToReturn = new WeeklyDiscountMovieDTO
+            if (weeklyDiscount == null)
             {
-                DiscountMovie = weeklyDiscount.WeeklyDiscount,
+                ChangeWeeklyDiscount();
+                weeklyDiscount = _movieRepository.GetWeeklyDiscountMovie();
+            }
+
+            var discountDTO = new DiscountDTO
+            {
+                DiscountMovie = weeklyDiscount.WeeklyDiscountMovie,
                 DiscountValue = weeklyDiscount.WeeklyDiscountValue
             };
-            return weeklyDiscountToReturn;
+
+            return discountDTO;
+        }
+
+        public void ChangeWeeklyDiscount()
+        {
+            var movies = _movieRepository.GetAllMovies();
+            if (movies == null)
+                throw new ListIsEmptyException();
+
+            _movieRepository.DrawNewWeeklyDiscount();
         }
     }
 }
